Change the 'App Path' so that the website plays nicely with GitHub Pages
I had to change the app-path so that this game doesn't link directly to my home page if someone clicks "Back to Game". Two solutions were present on this page: http://www.verdantrefuge.com/writing/2013/angularjs-changing-app-path-base-element/

The two solutions are "modifying the routes" and "use the <base> element". Both have their trade-offs.

I modified the routes directly because the annoyance of reading the modified routes sounds better than relying on the base tag (especially as commenters saw problems when using it, such as it forcing resets on older browsers). It also isn't that horrible to read.