Create a new political faction
May consider removing it though, considering that it is controversial. And yet, maybe I should keep it because it's controversial? The game takes place in a dystopian world, it's not meant to be a happy, funtime place. And people who get upset at censorship but are gleefully okay with violence...well, is that nice to have too?

At the same time, controversy is the last thing you want to have, and we live in our society.