Remove Firemen
The *apperance* of a conflict of interest is more dangerous than an actual conflict of interest. If you're a judge whose secretly supportive of EvilCorp, and you're very good at hiding your affiliation with EvilCorp, then you can get away with doing pro-EvilCorp actions without much of a peep. If you're not pro-EvilCorp, but if you're doing something that *appears* to show that you support EvilCorp, then you'll get (rightly) attacked for it.

It's the same thing here. The Firemen (as a concept) is not intended to take sides in political debates, but more of a reflection of a political issue taken to the extreme. Yet it *appears* as if I am taking a side (although which side that I take will depend on the reader in question). So it's probably better to just drop the faction.

On a side-note, I just realized that the Firemen aren't criminals at all. They aren't doing anything immoral. You have the freedom of speech to say anything, including the freedom to speak out *against* freedom of speech. The self-defeating nature of freedom notwithstanding, most Firemen are just average citizens living in a pretty hell-ish world and wanting easy solutions to complicated problems. It's hard to imagine a government cracking down on them without adding any more gimmicks on them, and honestly, death-threats, doxxing, and DDoSes (and any other extremist tactics that you could imagine the Firemen to use) would better fit the Human Flesh Search Engine.

The Firemen could be useful inspiration for a few event chains, but even then so what? Political correctness doesn't seem like something that would be offensive to the corporations (especially when they have a point), and language changes eventually. The 'future' Firemen may change media coverage in a few key areas, but offensive people will gleefully find new ways to say offensive things while still being "politically correct"...so the Firemen will win the battle but lose the war.

The Firemen's endless struggle to change societal norms could be a useful premise for another game, but it's not one that I want to make.