Upgrade HFSE from background game mechanic to non-playable faction
The reason I'm doing this is because I want to make the
game easier for me to make, and treating HFSE as a
faction would be easier than designing mechanics and fluff
that references it.

I also explored options in adding a traditional megacorp
back into the picture, but it looks to be very difficult.
I guess traditional megacorps just aren't that popular.